Revert "fixed compile bug and changed ulong to uint32_t"
This reverts commit 605b69b51e0d962ccf6a1072d4c18edc84e593ff.

+++ backend/include/Server/DatabaseComponent.hpp
@@ -15,13 +15,13 @@
 #include "Helper.h"
 
 struct r_CreateUser {
-    uint32_t puid;
+    ulong puid;
     std::string code;
 };
 
 struct r_LoginUser {
     bool successful;
-    uint32_t puid;
+    ulong puid;
     std::string sid;
     int error_code;
 };
@@ -44,10 +44,10 @@ public:
     Database(const mongocxx::uri &uri, const std::string &dbName, const std::string &collectionName);
 
     r_CreateUser createUser(std::string user_name, std::string user_email, std::string password);
-    bool verifyUser(uint32_t puid, std::string code);
+    bool verifyUser(ulong puid, std::string code);
     r_LoginUser loginUser(std::string email, std::string password);
-    bool deleteUser(uint32_t puid, std::string sid, std::string password);
-    oatpp::Object<UserDTO> getUser(uint32_t puid);
+    bool deleteUser(ulong puid, std::string sid, std::string password);
+    oatpp::Object<UserDTO> getUser(ulong puid);
     oatpp::Object<UserDTO> getUser(std::string field, std::string value);
     bool entryExists(std::string field, std::string value);
     void updateStringField( std::string filter_field, std::string filter_value, std::string field, std::string value);
